Secured Routes

+++ controllers/projectController.js
@@ -6,14 +6,14 @@ const Project = require("../models/projectModel")
 //@route GET /api/projects
 //@access public
 const getProjects = asyncHandler(async (req,res) => {
-    const projects = await Project.find();
+    const projects = await Project.find({ user_id: req.user.id });
     res.status(200).json(projects);
 });
 
 
 //@DESC Create a project
 //@route POST /api/projects
-//@access public
+//@access private
 const createProject = asyncHandler(async (req,res) => {
     console.log("The request body is:", req.body)
     const {name, manager, description, progress} = req.body;
@@ -25,7 +25,8 @@ const createProject = asyncHandler(async (req,res) => {
         name,
         manager,
         description,
-        progress
+        progress,
+        user_id: req.user.id,
     });
 
     res.status(200).json(project);
@@ -34,7 +35,7 @@ const createProject = asyncHandler(async (req,res) => {
 
 //@DESC Get a project
 //@route GET /api/projects/:id
-//@access public
+//@access private
 const getProject = asyncHandler(async (req,res) => {
     const project = await Project.findById(req.params.id);
     if(!project){
@@ -46,13 +47,18 @@ const getProject = asyncHandler(async (req,res) => {
 
 //@DESC Update a project
 //@route PUT /api/projects/:id
-//@access public
+//@access private
 const updateProject = asyncHandler(async (req,res) => {
     const project = await Project.findById(req.params.id);
     if (!project) {
         res.status(404);
         throw new Error("Contact not found");
-    }
+    };
+
+    if (project.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("You are not allowed to update this project");
+    };
 
     const updatedProject = await Project.findByIdAndUpdate(
         req.params.id,
@@ -66,13 +72,19 @@ const updateProject = asyncHandler(async (req,res) => {
 
 //@DESC Delete a project
 //@route DELETE /api/projects/:id
-//@access public
+//@access private
 const deleteProject = asyncHandler(async (req,res) => {
     const project = await Project.findById(req.params.id);
     if(!project){
         res.status(404);
         throw new Error("Project with Id not found!");
-    }
+    };
+
+    if (project.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("You are not allowed to delete this project");
+    };
+
     await Project.deleteOne({ _id: req.params.id });
     res.status(200).json(project);
 });
